feature(agent,pkg): turn connector into a Agent's mode
Move ShellHub Connector service into a is a new kind of ShellHub Agent
that turns Docker containers into ShellHub Devices. It instances a new
ShellHub Agent, in memory, for each container running, redirecting the
SSH IO, connections in general, and authentication credentials to the
container its file system.

> As most of the containers don't have passwords set for its users per default,
it rejects the connection every time for users without credentials.

The initial implementation has support for these kinds of connections:

- [x] Shell
- [x] Exec
- [x] Heredoc

To initialize the ShellHub Connector, enter the `agent/` directory, build and
run the agent's binary with *connector* sub command.

```sh
go build -ldflags "-X main.AgentVersion=latest" -o agent &&
SERVER_ADDRESS="http://localhost/" PRIVATE_KEYS="/tmp/shellhub/"
TENANT_ID="00000000-0000-4000-0000-000000000000" ./agent connector
```

To configure the ShellHub Connector, you can/must provide these environmental
variables.

- SERVER_ADDRESS (**required**) Set the ShellHub server address of the
  agent will use to connect

- PRIVATE_KEYS (**required**) Specify the path to store the devices/containers
  private keys. If not provided, the agent will generate a new one.

- TENANT_ID (**required**) Sets the account tenant ID used during communication
  to associate the devices to a specific tenant.

- KEEPALIVE_INTERVAL Determine the interval to send the keep alive message to
  the server. This has a direct impact on the bandwidth used by the device when
  in idle state.

As this implementation uses the Docker Client, you can override the environment
variables provided for Docker in order to change some behaviors, but the
ShellHub Connector doesn't guarantee its right operation with these changes.

Check [Docker Client documentation about
this](https://pkg.go.dev/github.com/docker/docker@v24.0.5+incompatible/client#
FromEnv) for more information.
