+++ .env
@@ -105,12 +105,6 @@ SHELLHUB_ASNYQ_GROUP_GRACE_PERIOD=1
 # the tasks immediately.
 SHELLHUB_ASYNQ_GROUP_MAX_SIZE=500
 
-# Enables container remote access
-SHELLHUB_CONNECTOR=false
-
-# This variable specifies the namespace to which containers will be added when the container remote access is enabled
-SHELLHUB_CONNECTOR_TENANT_ID=
-
 # Allows SSH to connect with an agent via a public key when the agent version is less than 0.6.0.
 # Agents 0.5.x or earlier do not validate the public key request and may panic.
 # Please refer to: https://github.com/shellhub-io/shellhub/issues/3453
