+++ .github/workflows/docker-publish.yml
@@ -12,7 +12,7 @@ jobs:
     strategy:
       fail-fast: true
       matrix:
-        project: [api, ssh, gateway, ui, cli, connector]
+        project: [api, ssh, gateway, ui, cli]
 
     runs-on: ubuntu-20.04
