+++ .github/workflows/qa.yml
@@ -16,7 +16,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        project: [api, agent, pkg, ssh, ui, cli, connector]
+        project: [api, agent, pkg, ssh, ui, cli]
         include:
           - project: api
             extra_args: ""
@@ -34,9 +34,6 @@ jobs:
             extra_args: ""
             lint_args: ""
           - project: cli
-            extra_args: ""
-            lint_args: ""
-          - project: connector
             extra_args: ""
             lint_args: ""
           - project: tests
@@ -62,7 +59,6 @@ jobs:
               - 'pkg/**'
               - 'ssh/**'
               - 'cli/**'
-              - 'connector/**'
               - 'tests/**'
 
       - name: Set up Go 1.x [Go]
