--- connector/entrypoint-dev.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-refresh run
