--- connector/refresh.yml
@@ -1,14 +1,0 @@
-app_root: /go/src/github.com/shellhub-io/shellhub
-ignored_folders:
-  - vendor
-included_extensions:
-  - .go
-build_target_path: ""
-build_path: /go/src/github.com/shellhub-io/shellhub/connector
-build_flags: ["-tags", "docker", "-ldflags", "-X main.ConnectorVersion=latest"]
-build_delay: 200ns
-binary_name: connector
-command_flags: ["docker"]
-command_env: []
-enable_colors: true
-log_name: ""
