--- docker-compose.connector.yml
@@ -1,20 +1,0 @@
-services:
-  connector:
-    image: shellhubio/connector:${SHELLHUB_VERSION}
-    restart: unless-stopped
-    network_mode: host
-    environment:
-      - SHELLHUB_SERVER_ADDRESS=http://${SHELLHUB_DOMAIN}/
-      - SHELLHUB_PRIVATE_KEYS=/var/run/connector/keys
-      - SHELLHUB_TENANT_ID=${SHELLHUB_CONNECTOR_TENANT_ID}
-      - SHELLHUB_LOG_LEVEL=${SHELLHUB_LOG_LEVEL}
-      - SHELLHUB_LOG_FORMAT=${SHELLHUB_LOG_FORMAT}
-    volumes:
-      - /var/run/docker.sock:/var/run/docker.sock
-      - "connector_keys:/var/run/connector/keys"
-    depends_on:
-      - api
-      - ssh
-      - gateway
-volumes:
-  connector_keys:
